Move logic out to a concern
