+++ decidim-admin/app/controllers/decidim/admin/application_controller.rb
@@ -7,6 +7,7 @@ module Decidim
       include NeedsAuthorization
       include FormFactory
       include LocaleSwitcher
+      include PayloadInfo
       helper Decidim::DecidimFormHelper
       helper Decidim::ReplaceButtonsHelper
       helper Decidim::OrganizationScopesHelper
@@ -24,18 +25,6 @@ module Decidim
       # since the gem expects the ability class to be in the root namespace.
       def current_ability_klass
         Decidim::Admin::Abilities::Base
-      end
-
-      def append_info_to_payload(payload)
-        super
-        payload[:user_id] = current_user.id
-        payload[:organization_id] = current_organization.id
-        payload[:app] = current_organization.name
-        payload[:remote_ip] = request.remote_ip
-        payload[:referer] = request.referer.to_s
-        payload[:request_id] = request.uuid
-        payload[:user_agent] = request.user_agent
-        payload[:xhr] = request.xhr? ? 'true' : 'false'
       end
     end
   end
