+++ decidim-core/app/controllers/concerns/decidim/payload_info.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require "active_support/concern"
+
+module Decidim
+  module PayloadInfo
+    extend ActiveSupport::Concern
+
+    included do
+      def append_info_to_payload(payload)
+        super
+        payload[:user_id] = current_user.id
+        payload[:organization_id] = current_organization.id
+        payload[:app] = current_organization.name
+        payload[:remote_ip] = request.remote_ip
+        payload[:referer] = request.referer.to_s
+        payload[:request_id] = request.uuid
+        payload[:user_agent] = request.user_agent
+        payload[:xhr] = request.xhr? ? 'true' : 'false'
+      end
+    end
+  end
+end
