+++ decidim-core/app/controllers/decidim/application_controller.rb
@@ -5,6 +5,7 @@ module Decidim
     include Decidim::NeedsOrganization
     include Decidim::LocaleSwitcher
     include NeedsAuthorization
+    include PayloadInfo
 
     helper Decidim::MetaTagsHelper
     helper Decidim::DecidimFormHelper
@@ -46,18 +47,6 @@ module Decidim
 
     def redirect_to_404
       raise ActionController::RoutingError, "Not Found"
-    end
-
-    def append_info_to_payload(payload)
-      super
-      payload[:user_id] = current_user.id
-      payload[:organization_id] = current_organization.id
-      payload[:app] = current_organization.name
-      payload[:remote_ip] = request.remote_ip
-      payload[:referer] = request.referer.to_s
-      payload[:request_id] = request.uuid
-      payload[:user_agent] = request.user_agent
-      payload[:xhr] = request.xhr? ? 'true' : 'false'
     end
   end
 end
